Project Update (Week 3)
Now, the blog page are updated using some array functions. There are also single post page for every post on the blog page. Each page have their own slug to differentiate every single page. Model are also created to represent the database. I created the model for post page so the web route can call it to display the data whenever the user click on the page link. 404 error page will be displayed if the the link is wrong or unavailable etc.