Project Update (Week 4)
This update is about the creation of the database. My data is stored using TablePlus application. For this case, I was storing the article posts data including the title, author, slug & body. I added the new post (data) to the database using the eloquent ORM technique. I type 'php artisan tinker' to the terminal and then I typed ' App/model/Post::create([ ]); ' to add the article post data and it will added to the posts database inside the TablePlus automatically.